requirements.txt
appdirs==1.4.4
appengine-mailer==0.1
argcomplete==2.0.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.14.1
autopep8==2.0.1
beautifulsoup4==4.11.2
cachetools==5.3.0
catkin-pkg==0.5.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
dateparser==1.1.7
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
dj-rest-auth==2.2.4
django-rest-auth==0.9.5
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-filter==22.1
django-jsonfield==1.4.1
django-otp==1.1.4
django-pandas==0.6.6
django-perms==2.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-rocket-engine==0.0.11
django-sendgrid-v5==1.2.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
docopt==0.6.2
docutils==0.19
et-xmlfile==1.1.0
filelock==3.9.0
future==0.18.3
google==3.0.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
isort==5.12.0
itypes==1.2.0
jdcal==1.4.1
jeepney==0.8.0
Jinja2==3.1.2
jsonfield==3.1.0
keyring==23.13.1
keyrings.alt==4.2.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
moment==0.12.1
oauth==1.0.1
oauthlib==3.2.2
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.1.0
pandoradep==0.3.0
pbr==5.5.1
pexpect==4.8.0
pipreqs==0.4.12
protobuf
psycopg==3.1.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
pylint==2.16.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
python-dateutil==2.8.2
python-debian==0.1.39
python-gnupg==0.5.0
python-http-client==3.3.7
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scour==0.38.2
SecretStorage==3.3.3
sendgrid==6.9.7
simplejson==3.18.3
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.3
testresources==2.0.1
times==0.7
typing_extensions==4.5.0
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
yarg==0.1.9
zope.interface==5.5.2